refactoring how i'll handle markdowns
