+++ lua/core/plugins/markdown/init.lua
@@ -1,0 +1,4 @@
+return {
+  require("core.plugins.markdown.render-markdown"),
+  require("core.plugins.markdown.autolist"),
+}
